updated migrations file

+++ database/migrations/2017_03_02_145003_create_table_permission_role.php
@@ -13,14 +13,17 @@ class CreateTablePermissionRole extends Migration
      */
     public function up()
     {
-        Schema::create('permission_role', function (Blueprint $table) {
-            $table->increments('id');
-            $table->integer('permission_id')->unsigned();
-            $table->integer('role_id')->unsigned();
-            $table->foreign('permission_id')->references('id')->on('permissions')->onUpdate('cascade')->onDelete('cascade');
-            $table->foreign('role_id')->references('id')->on('roles')->onUpdate('cascade')->onDelete('cascade');
-            $table->timestamps();
-        });
+        if(!Schema::hasTable('permission_role'))
+        {
+            Schema::create('permission_role', function (Blueprint $table) {
+                $table->increments('id');
+                $table->integer('permission_id')->unsigned();
+                $table->integer('role_id')->unsigned();
+                $table->foreign('permission_id')->references('id')->on('permissions')->onUpdate('cascade')->onDelete('cascade');
+                $table->foreign('role_id')->references('id')->on('roles')->onUpdate('cascade')->onDelete('cascade');
+                $table->timestamps();
+            });
+        }
     }
 
     /**
